+++ AlphaFS/Filesystem/CopyMoveProgressRoutine.cs
@@ -19,8 +19,11 @@
  *  THE SOFTWARE. 
  */
 
+using System;
+
 namespace Alphaleonis.Win32.Filesystem
 {
    /// <summary>Callback used by CopyFileXxx and MoveFileXxx to report progress about the copy/move operation.</summary>
+   [CLSCompliant(false)]
    public delegate CopyMoveProgressResult CopyMoveProgressRoutine(long totalFileSize, long totalBytesTransferred, long streamSize, long streamBytesTransferred, uint streamNumber, CopyMoveProgressCallbackReason callbackReason, object userData);
 }