+++ AlphaFS/Filesystem/DiskSpaceInfo.cs
@@ -265,6 +265,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region FreeBytesAvailable
 
       /// <summary>The total number of free bytes on a disk that are available to the user who is associated with the calling thread.</summary>
+      [CLSCompliant(false)]
       public ulong FreeBytesAvailable { get; private set; }
 
       #endregion // FreeBytesAvailable
@@ -272,6 +273,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region TotalNumberOfBytes
 
       /// <summary>The total number of bytes on a disk that are available to the user who is associated with the calling thread.</summary>
+      [CLSCompliant(false)]
       public ulong TotalNumberOfBytes { get; private set; }
 
       #endregion // TotalNumberOfBytes
@@ -279,6 +281,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region TotalNumberOfFreeBytes
 
       /// <summary>The total number of free bytes on a disk.</summary>
+      [CLSCompliant(false)]
       public ulong TotalNumberOfFreeBytes { get; private set; }
 
       #endregion // TotalNumberOfFreeBytes
@@ -287,6 +290,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region BytesPerSector
 
       /// <summary>The number of bytes per sector.</summary>
+      [CLSCompliant(false)]
       public uint BytesPerSector { get; private set; }
 
       #endregion // BytesPerSector
@@ -294,6 +298,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region NumberOfFreeClusters
 
       /// <summary>The total number of free clusters on the disk that are available to the user who is associated with the calling thread.</summary>
+      [CLSCompliant(false)]
       public uint NumberOfFreeClusters { get; private set; }
 
       #endregion // NumberOfFreeClusters
@@ -301,6 +306,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region SectorsPerCluster
 
       /// <summary>The number of sectors per cluster.</summary>
+      [CLSCompliant(false)]
       public uint SectorsPerCluster { get; private set; }
 
       #endregion // SectorsPerCluster
@@ -310,6 +316,7 @@ namespace Alphaleonis.Win32.Filesystem
       /// <summary>The total number of clusters on the disk that are available to the user who is associated with the calling thread.
       /// If per-user disk quotas are in use, this value may be less than the total number of clusters on the disk.
       /// </summary>
+      [CLSCompliant(false)]
       public uint TotalNumberOfClusters { get; private set; }
 
       #endregion // TotalNumberOfClusters