+++ AlphaFS/Filesystem/FileIdBothDirectoryInfo.cs
@@ -86,6 +86,7 @@ namespace Alphaleonis.Win32.Filesystem
       #region EaSize
 
       /// <summary>The size of the extended attributes for the file.</summary>
+      [CLSCompliant(false)]
       public uint ExtendedAttributesSize { get; set; }
 
       #endregion // EaSize
@@ -119,6 +120,7 @@ namespace Alphaleonis.Win32.Filesystem
       /// <summary>The byte offset of the file within the parent directory. This member is undefined for file systems, such as NTFS,
       /// in which the position of a file within the parent directory is not fixed and can be changed at any time to maintain sort order.
       /// </summary>
+      [CLSCompliant(false)]
       public uint FileIndex { get; set; }
 
       #endregion // FileIndex