+++ AlphaFS/Filesystem/FileSystemEntryInfo.cs
@@ -201,6 +201,7 @@ namespace Alphaleonis.Win32.Filesystem
 
       /// <summary>Gets the reparse point tag of this entry.</summary>
       /// <value>The reparse point tag of this entry.</value>
+      [CLSCompliant(false)]
       public ReparsePointTag ReparsePointTag
       {
          get { return IsReparsePoint ? _win32FindData.Reserved0 : ReparsePointTag.None; }