+++ AlphaFS/Network/DfsInfo.cs
@@ -146,6 +146,7 @@ namespace Alphaleonis.Win32.Network
       #region Timeout
 
       /// <summary>Specifies the time-out, in seconds, of the DFS root or link.</summary>
+      [CLSCompliant(false)]
       public ulong Timeout { get; internal set; }
 
       #endregion // Timeout